+++ lib/marskal.rb
@@ -1,5 +1,15 @@
 require "marskal/version"
+require "marskal/core_ext"
 
+require 'active_support/dependencies/autoload'
+require 'active_support/core_ext/numeric'
+require 'active_support/core_ext/date'
+
+##
+# Marskal.
+#
+# This gem contains the basic extensions to ruby methods and custom utilities and other methods and functionality that
+# act as support and provide inheritance for other layers of 'marskal-???' gems and applications
+#
 module Marskal
-  # Your code goes here...
 end