+++ CHANGELOG.md
@@ -1,7 +1,20 @@
 # Changelog
 
+#### 0.1.1.wip - MAU `2016-12-07`  
+**Added core_ext and implemented documentation options for testing**
+
+* Added in core_ext, take from gem marskal-core-extensions
+* Created test documents for one of the Array methods
+* Added a rake utiltiy to generate docs:
+    * Rdoc default
+    * Hanna Nouveau
+    * Yard
+    
+---
+
 #### 0.1.0.wip - MAU `2016-12-07`  
 **Initial Version**
+
 * Stubbed in the basic file structures
 * Make sure dummy tests are running properly